requirements.txt
flask==2.3.3
werkzeug==2.3.7
pillow==9.5.0
pyautogui==0.9.54
python-engineio==4.2.1
python-socketio==5.4.0
numpy==1.24.3
opencv-python-headless==4.8.0.74
